% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{add_constraint}
\alias{add_constraint}
\title{Ajoute une contrainte à une grille}
\usage{
add_constraint(grid, row, col, value)
}
\arguments{
\item{grid}{Objet SlitherlinkGrid}

\item{row}{Ligne de la case}

\item{col}{Colonne de la case}

\item{value}{Valeur de la contrainte (0, 1, 2, ou 3)}
}
\value{
La grille modifiée
}
\description{
Ajoute une contrainte à une grille
}
\examples{
grid <- create_grid(5, 5)
grid <- add_constraint(grid, 2, 2, 3)

}
