% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{backtrack_solve}
\alias{backtrack_solve}
\title{Algorithme de backtracking}
\usage{
backtrack_solve(grid, possible_edges, edge_index, max_iter, counter_env)
}
\description{
Algorithme de backtracking
}
\keyword{internal}
