% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_constraints}
\alias{check_constraints}
\title{Vérifie que toutes les contraintes de la case sont respectées}
\usage{
check_constraints(grid)
}
\arguments{
\item{grid}{Objet SlitherlinkGrid}
}
\value{
TRUE si toutes les contraintes sont respectées, FALSE sinon
}
\description{
Pour chaque case avec un chiffre, vérifie que le nombre d'arêtes
tracées autour de cette case correspond exactement au chiffre.
}
