% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{check_degree}
\alias{check_degree}
\title{Vérifier le degré de chaque sommet}
\usage{
check_degree(grid)
}
\arguments{
\item{grid}{Un objet de classe \code{slitherlink}.}
}
\value{
Une liste : \code{ok} (logical), \code{errors} (dataframe des sommets en erreur).
}
\description{
À chaque point d'intersection, le nombre de segments tracés
doit être exactement 0 ou 2 (jamais 1 ni 3).
}
