% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{create_grid}
\alias{create_grid}
\title{Crée une nouvelle grille Slitherlink}
\usage{
create_grid(width = 5, height = 5)
}
\arguments{
\item{width}{Largeur de la grille (colonnes)}

\item{height}{Hauteur de la grille (lignes)}
}
\value{
Un objet SlitherlinkGrid
}
\description{
Crée une nouvelle grille Slitherlink
}
\examples{
grid <- create_grid(5, 5)
grid$add_constraint(2, 2, 3)

}
