% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw_grid_ggplot}
\alias{draw_grid_ggplot}
\title{Dessine une grille avec ggplot2}
\usage{
draw_grid_ggplot(grid)
}
\arguments{
\item{grid}{Objet SlitherlinkGrid}
}
\value{
Un objet ggplot
}
\description{
Dessine une grille avec ggplot2
}
