% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{generate_random_loop}
\alias{generate_random_loop}
\title{Générer une boucle complexe aléatoire}
\usage{
generate_random_loop(n, m)
}
\arguments{
\item{n, m}{Dimensions de la grille}
}
\value{
Une liste (h_edges, v_edges)
}
\description{
Générer une boucle complexe aléatoire
}
