% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{get_all_possible_edges}
\alias{get_all_possible_edges}
\title{Génère toutes les arêtes possibles (L'ULTIME OPTIMISATION)}
\usage{
get_all_possible_edges(grid)
}
\arguments{
\item{grid}{SlitherlinkGrid object}
}
\value{
Liste d'arêtes
}
\description{
Combine les DEUX meilleures stratégies :
1. Exclure les arêtes des cases '0' (réduit l'espace de recherche).
2. Ordre géographique (permet de déclencher l'élagage très tôt).
}
\keyword{internal}
