% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grid_statistics}
\alias{grid_statistics}
\title{Calcule des statistiques sur une grille}
\usage{
grid_statistics(grid)
}
\arguments{
\item{grid}{Objet SlitherlinkGrid}
}
\value{
Liste de statistiques
}
\description{
Calcule des statistiques sur une grille
}
