% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grid_to_json}
\alias{grid_to_json}
\title{Convertit une grille en format JSON}
\usage{
grid_to_json(grid)
}
\arguments{
\item{grid}{Objet SlitherlinkGrid}
}
\value{
Chaîne JSON
}
\description{
Convertit une grille en format JSON
}
