% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{init_db}
\alias{init_db}
\title{Initialiser la base de données}
\usage{
init_db(db_path = "leaderboard.db")
}
\arguments{
\item{db_path}{Chemin vers le fichier de base de données.}
}
\description{
Initialiser la base de données
}
