% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{json_to_grid}
\alias{json_to_grid}
\title{Convertit du JSON en grille}
\usage{
json_to_grid(json_str)
}
\arguments{
\item{json_str}{Chaîne JSON}
}
\value{
Objet SlitherlinkGrid
}
\description{
Convertit du JSON en grille
}
