% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_management.R
\name{new_slitherlink}
\alias{new_slitherlink}
\title{Création d'une structure Slitherlink rigoureuse}
\usage{
new_slitherlink(indices)
}
\arguments{
\item{indices}{Une matrice d'entiers (0-3 ou NA) représentant les contraintes.}
}
\value{
Un objet de classe \code{slitherlink}.
}
\description{
Création d'une structure Slitherlink rigoureuse
}
