% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_grid}
\alias{plot_grid}
\title{Affiche une grille dans la console (mode texte)}
\usage{
plot_grid(grid)
}
\arguments{
\item{grid}{Objet SlitherlinkGrid}
}
\description{
Affiche une grille dans la console (mode texte)
}
