% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_plot.R
\name{plot_slitherlink}
\alias{plot_slitherlink}
\title{Visualisation de la grille Slitherlink}
\usage{
plot_slitherlink(grid)
}
\arguments{
\item{grid}{Un objet de classe \code{slitherlink}}
}
\description{
Visualisation de la grille Slitherlink
}
