% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_solution}
\alias{plot_solution}
\title{Affiche une solution dans la console}
\usage{
plot_solution(grid)
}
\arguments{
\item{grid}{Objet SlitherlinkGrid avec solution}
}
\description{
Affiche une solution dans la console
}
