% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_statistics}
\alias{print_statistics}
\title{Affiche les statistiques d'une grille}
\usage{
print_statistics(grid)
}
\arguments{
\item{grid}{Objet SlitherlinkGrid}
}
\description{
Affiche les statistiques d'une grille
}
