% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_game.R
\name{run_slitherlink}
\alias{run_slitherlink}
\title{Lancer l'application Slitherlink}
\usage{
run_slitherlink()
}
\description{
Cette fonction trouve l'application Shiny installée dans le package
et la lance.
}
