% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_slitherlink_app}
\alias{run_slitherlink_app}
\title{Lance l'application Shiny}
\usage{
run_slitherlink_app(grid = NULL)
}
\arguments{
\item{grid}{Grille optionnelle à charger au démarrage}
}
\description{
Lance l'application Shiny
}
