% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{save_score}
\alias{save_score}
\title{Sauvegarder le score}
\usage{
save_score(pseudo, temps, diff, db_path = "leaderboard.db")
}
\arguments{
\item{pseudo}{Pseudo du joueur}

\item{temps}{Temps en secondes}

\item{diff}{Niveau de difficulté}

\item{db_path}{Chemin vers le fichier de base de données.}
}
\description{
Sauvegarder le score
}
