% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{SlitherlinkGrid}
\alias{SlitherlinkGrid}
\title{Classe SlitherlinkGrid}
\description{
Représente une grille de jeu Slitherlink.
Une grille est composée de points et d'arêtes potentielles.
Les contraintes (chiffres) sont stockées dans les cases.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{width}}{Largeur de la grille (nombre de colonnes de cases)}

\item{\code{height}}{Hauteur de la grille (nombre de lignes de cases)}

\item{\code{constraints}}{Matrice des contraintes (chiffres dans les cases)}

\item{\code{edges}}{Liste des arêtes tracées (solution actuelle)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SlitherlinkGrid-new}{\code{SlitherlinkGrid$new()}}
\item \href{#method-SlitherlinkGrid-add_constraint}{\code{SlitherlinkGrid$add_constraint()}}
\item \href{#method-SlitherlinkGrid-add_edge}{\code{SlitherlinkGrid$add_edge()}}
\item \href{#method-SlitherlinkGrid-remove_edge}{\code{SlitherlinkGrid$remove_edge()}}
\item \href{#method-SlitherlinkGrid-are_adjacent}{\code{SlitherlinkGrid$are_adjacent()}}
\item \href{#method-SlitherlinkGrid-has_edge}{\code{SlitherlinkGrid$has_edge()}}
\item \href{#method-SlitherlinkGrid-count_edges_around_cell}{\code{SlitherlinkGrid$count_edges_around_cell()}}
\item \href{#method-SlitherlinkGrid-print}{\code{SlitherlinkGrid$print()}}
\item \href{#method-SlitherlinkGrid-clone}{\code{SlitherlinkGrid$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlitherlinkGrid-new"></a>}}
\if{latex}{\out{\hypertarget{method-SlitherlinkGrid-new}{}}}
\subsection{Method \code{new()}}{
Initialise une nouvelle grille Slitherlink
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlitherlinkGrid$new(width = 5, height = 5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{width}}{Largeur (colonnes de cases)}

\item{\code{height}}{Hauteur (lignes de cases)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Un nouvel objet SlitherlinkGrid
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlitherlinkGrid-add_constraint"></a>}}
\if{latex}{\out{\hypertarget{method-SlitherlinkGrid-add_constraint}{}}}
\subsection{Method \code{add_constraint()}}{
Ajoute une contrainte (chiffre) dans une case
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlitherlinkGrid$add_constraint(row, col, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{row}}{Ligne de la case (1 à height)}

\item{\code{col}}{Colonne de la case (1 à width)}

\item{\code{value}}{Valeur de la contrainte (0, 1, 2, ou 3)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlitherlinkGrid-add_edge"></a>}}
\if{latex}{\out{\hypertarget{method-SlitherlinkGrid-add_edge}{}}}
\subsection{Method \code{add_edge()}}{
Ajoute une arête (ligne) à la solution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlitherlinkGrid$add_edge(from, to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{from}}{Point de départ (vecteur c(row, col))}

\item{\code{to}}{Point d'arrivée (vecteur c(row, col))}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlitherlinkGrid-remove_edge"></a>}}
\if{latex}{\out{\hypertarget{method-SlitherlinkGrid-remove_edge}{}}}
\subsection{Method \code{remove_edge()}}{
Retire une arête de la solution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlitherlinkGrid$remove_edge(from, to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{from}}{Point de départ}

\item{\code{to}}{Point d'arrivée}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlitherlinkGrid-are_adjacent"></a>}}
\if{latex}{\out{\hypertarget{method-SlitherlinkGrid-are_adjacent}{}}}
\subsection{Method \code{are_adjacent()}}{
Vérifie si deux points sont adjacents
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlitherlinkGrid$are_adjacent(p1, p2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p1}}{Premier point c(row, col)}

\item{\code{p2}}{Deuxième point c(row, col)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE si adjacents, FALSE sinon
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlitherlinkGrid-has_edge"></a>}}
\if{latex}{\out{\hypertarget{method-SlitherlinkGrid-has_edge}{}}}
\subsection{Method \code{has_edge()}}{
Vérifie si une arête existe déjà
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlitherlinkGrid$has_edge(from, to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{from}}{Point de départ}

\item{\code{to}}{Point d'arrivée}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE si l'arête existe, FALSE sinon
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlitherlinkGrid-count_edges_around_cell"></a>}}
\if{latex}{\out{\hypertarget{method-SlitherlinkGrid-count_edges_around_cell}{}}}
\subsection{Method \code{count_edges_around_cell()}}{
Compte le nombre d'arêtes autour d'une case
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlitherlinkGrid$count_edges_around_cell(row, col)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{row}}{Ligne de la case}

\item{\code{col}}{Colonne de la case}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nombre d'arêtes tracées autour de la case
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlitherlinkGrid-print"></a>}}
\if{latex}{\out{\hypertarget{method-SlitherlinkGrid-print}{}}}
\subsection{Method \code{print()}}{
Affiche la grille (pour débogage)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlitherlinkGrid$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SlitherlinkGrid-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SlitherlinkGrid-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SlitherlinkGrid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
