% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_interface.R
\name{solve_game}
\alias{solve_game}
\title{Résoudre une grille Slitherlink}
\usage{
solve_game(grid)
}
\arguments{
\item{grid}{Un objet de classe \code{slitherlink}.}
}
\value{
Une liste contenant \code{success} (booléen) et \code{grid} (l'objet mis à jour).
}
\description{
Résoudre une grille Slitherlink
}
