% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solve_slitherlink}
\alias{solve_slitherlink}
\title{Résoudre un puzzle Slitherlink via ILP}
\usage{
solve_slitherlink(grid)
}
\arguments{
\item{grid}{Un objet de classe \code{slitherlink}}
}
\value{
La grille résolue ou NULL
}
\description{
Résoudre un puzzle Slitherlink via ILP
}
