% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve_slitherlink_cpp}
\alias{solve_slitherlink_cpp}
\title{Solveur Slitherlink Optimisé C++}
\usage{
solve_slitherlink_cpp(h_edges, v_edges, indices)
}
\arguments{
\item{h_edges}{Matrice segments horizontaux}

\item{v_edges}{Matrice segments verticaux}

\item{indices}{Matrice des chiffres contraintes}
}
\description{
Solveur Slitherlink Optimisé C++
}
