% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_management.R
\name{toggle_h_edge}
\alias{toggle_h_edge}
\title{Basculer l'état d'un segment horizontal}
\usage{
toggle_h_edge(grid, r, c)
}
\arguments{
\item{grid}{Un objet de classe \code{slitherlink}.}

\item{r}{Indice de ligne   (1 à n+1).}

\item{c}{Indice de colonne (1 à m).}
}
\value{
La grille modifiée.
}
\description{
Basculer l'état d'un segment horizontal
}
