% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_management.R
\name{toggle_v_edge}
\alias{toggle_v_edge}
\title{Basculer l'état d'un segment vertical}
\usage{
toggle_v_edge(grid, r, c)
}
\arguments{
\item{grid}{Un objet de classe \code{slitherlink}.}

\item{r}{Indice de ligne   (1 à n).}

\item{c}{Indice de colonne (1 à m+1).}
}
\value{
La grille modifiée.
}
\description{
Basculer l'état d'un segment vertical
}
