% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{generate_puzzle}
\alias{generate_puzzle}
\title{Générer un puzzle Slitherlink}
\usage{
generate_puzzle(difficulty = "moyen", n = 5, m = 5)
}
\arguments{
\item{difficulty}{Niveau de difficulté ("facile", "moyen" ou "difficile")}

\item{n}{Nombre de lignes de cases (défaut 5)}

\item{m}{Nombre de colonnes de cases (défaut 5)}
}
\description{
Générer un puzzle Slitherlink
}
