% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{generate_puzzle}
\alias{generate_puzzle}
\title{Générer un puzzle Slitherlink depuis une boucle avec indices masqués}
\usage{
generate_puzzle(h_sol, v_sol, difficulty = "moyen")
}
\arguments{
\item{h_sol}{Matrice h_edges de la solution}

\item{v_sol}{Matrice v_edges de la solution}

\item{difficulty}{"facile" (peu de NA), "moyen", "difficile" (beaucoup de NA)}
}
\value{
Une liste avec grid (indices partiels) et solution
}
\description{
Générer un puzzle Slitherlink depuis une boucle avec indices masqués
}
