% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{indices_from_loop}
\alias{indices_from_loop}
\title{Calculer les indices depuis une boucle}
\usage{
indices_from_loop(h, v)
}
\arguments{
\item{h, v}{Matrices de segments}
}
\description{
Calculer les indices depuis une boucle
}
