% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{is_unique_solution}
\alias{is_unique_solution}
\title{Verifier qu'un puzzle a une solution unique}
\usage{
is_unique_solution(grid, sol_h, sol_v)
}
\arguments{
\item{grid}{Un objet de classe \code{slitherlink}}

\item{sol_h}{Matrice h_edges de la solution}

\item{sol_v}{Matrice v_edges de la solution}
}
\value{
TRUE si la solution est unique, FALSE sinon
}
\description{
Verifier qu'un puzzle a une solution unique
}
