% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solve_puzzle}
\alias{solve_puzzle}
\title{Résout un puzzle Slitherlink}
\usage{
solve_puzzle(grid, max_iterations = 1e+07)
}
\arguments{
\item{grid}{Objet SlitherlinkGrid avec les contraintes}

\item{max_iterations}{Nombre maximum d'itérations}
}
\value{
Une grille avec la solution, ou NULL
}
\description{
Résout un puzzle Slitherlink
}
